move more logic out of instrumentdata, add tests for logic moved
