+++ jest.config.ts
@@ -19,6 +19,9 @@ const config: Config = {
     "!app/components/JenkinsJobs.tsx", // relies on an external fetch
     "!app/components/ShowHideBeamInfo.tsx", // relies on an external image
     "!app/components/InstrumentData.tsx", // relies on websocket
+    "!app/wall/page.tsx", // relies on JenkinsJobs
+    "!app/instruments/page.tsx", // no logic here
+    "!app/instrument/page.tsx", // no logic here - all in instrumentdata / instrument.ts
   ],
 };
